Setup PIRS infrastructure and a demo button in the user talk page for testing
This patch is to finalize setting up the IncidentReporting
extension from BoilerPlate. The purpose of this patch is to
add a minimal report button to the user talk page namespace.
This button is not part of the design requirements, but rather
a demo that all is setup and configured correctly.

This button will be replaced by the designs as specified in
https://www.figma.com/file/vbxjEvCFNsaJyT5m45kjSX/Private-Incident-Reporting-System?type=design&mode=design&t=V7gwGiaQOZDjmSrv-0

Bug: T337533
Change-Id: I2e166baab850061db7997034a43e072481985086